MoveAlongPathEffect separated from MoveEffect
