+++ packages/flame/lib/effects.dart
@@ -11,6 +11,7 @@ export 'src/effects/controllers/reverse_curved_effect_controller.dart';
 export 'src/effects/controllers/reverse_linear_effect_controller.dart';
 export 'src/effects/controllers/sequence_effect_controller.dart';
 export 'src/effects/effect.dart';
+export 'src/effects/move_along_path_effect.dart';
 export 'src/effects/move_effect.dart';
 export 'src/effects/opacity_effect.dart';
 export 'src/effects/remove_effect.dart';
