+++ packages/flame/lib/src/effects/move_effect.dart
@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 import 'package:vector_math/vector_math_64.dart';
 
 import 'controllers/effect_controller.dart';
@@ -16,11 +14,6 @@ import 'transform2d_effect.dart';
 ///   - [MoveEffect.to] will move the target in a straight line to the specified
 ///     coordinates;
 ///
-///   - [MoveEffect.along] will move the target along the specified path, which
-///     may contain curved segments, but must be simply-connected. The `path`
-///     argument is taken as relative to the target's position at the start of
-///     the effect.
-///
 /// This effect applies incremental changes to the component's position, and
 /// requires that any other effect or update logic applied to the same component
 /// also used incremental updates.
@@ -31,9 +24,6 @@ class MoveEffect extends Transform2DEffect {
 
   factory MoveEffect.to(Vector2 destination, EffectController controller) =>
       _MoveToEffect(destination, controller);
-
-  factory MoveEffect.along(Path path, EffectController controller) =>
-      _MoveAlongPathEffect(path, controller);
 
   Vector2 _offset;
 
@@ -56,43 +46,5 @@ class _MoveToEffect extends MoveEffect {
   @override
   void onStart() {
     _offset = _destination - target.position;
-  }
-}
-
-/// Implementation class for [MoveEffect.along].
-class _MoveAlongPathEffect extends MoveEffect {
-  _MoveAlongPathEffect(Path path, EffectController controller)
-      : _lastPosition = Vector2.zero(),
-        super.by(Vector2.zero(), controller) {
-    final metrics = path.computeMetrics().toList();
-    if (metrics.length != 1) {
-      throw ArgumentError(
-        'Only single-contour paths are allowed in MoveEffect.along',
-      );
-    }
-    _pathMetric = metrics[0];
-    _pathLength = _pathMetric.length;
-    assert(_pathLength > 0);
-  }
-
-  late final PathMetric _pathMetric;
-  late final double _pathLength;
-  late Vector2 _lastPosition;
-
-  @override
-  void apply(double progress) {
-    final distance = progress * _pathLength;
-    final tangent = _pathMetric.getTangentForOffset(distance)!;
-    final offset = tangent.position;
-    target.position.x += offset.dx - _lastPosition.x;
-    target.position.y += offset.dy - _lastPosition.y;
-    _lastPosition.setValues(offset.dx, offset.dy);
-    super.apply(progress);
-  }
-
-  @override
-  void reset() {
-    super.reset();
-    _lastPosition = Vector2.zero();
   }
 }
